+++ R/woven-package.R
@@ -5,6 +5,8 @@
 #' @importFrom RANN nn2
 #' @importFrom cluster silhouette
 #' @importFrom MatrixGenerics colMedians
-#' @importFrom parallel detectCores mclapply
+#' @importFrom parallel detectCores
+#' @importFrom BiocParallel bplapply SerialParam MulticoreParam SnowParam
+#' @importFrom methods is new setValidity show
 #' @importFrom stats coef cor dist kmeans lm median model.matrix rnorm var
 NULL